% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{normalizeGeneLength}
\alias{normalizeGeneLength}
\title{Normalize for gene length}
\usage{
normalizeGeneLength(...)
}
\arguments{
\item{...}{...}
}
\description{
Normalize for gene length using the output of transcript abundance estimators
}
\details{
This function is deprecated and moved to a new general purpose package,
tximport, which will be added to Bioconductor.
}
