% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotPCA}
\alias{plotPCA}
\title{Sample PCA plot from variance-stabilized data}
\usage{
plotPCA(x, intgroup = "condition", ntop = 500, col)
}
\arguments{
\item{x}{a SummarizedExperiment, with data in \code{assay(x)},
produced for example by either \code{\link{varianceStabilizingTransformation}}
or \code{\link{rlogTransformation}}}

\item{intgroup}{a character vector of names in \code{colData(x)} to use for grouping}

\item{ntop}{number of top genes to use for principal components, selected by highest
row variance}

\item{col}{a vector of colors for each level of intgroup}
}
\value{
A \code{trellis} object.
}
\description{
This plot helps to check for batch effects and the like.
}
\note{
See the vignette for an example of variance stabilization and PCA plots.
}
\examples{
dds = makeExampleDESeqDataSet(betaSD=1)
vsd = varianceStabilizingTransformation(dds)
p = plotPCA(vsd)
print(p)

## Add text labels (for presentation graphics, consider additional
## layout operations that avoid overplotting, such as the FField package on CRAN)
names = colData(vsd)$sample

p = update(p, panel = function(x, y, ...) {
      lattice::panel.xyplot(x, y, ...);
      lattice::ltext(x=x, y=y, labels=names, pos=1, offset=1, cex=0.8)
    })
print(p)
}
\author{
Wolfgang Huber
}

