% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{priorInfo}
\alias{priorInfo}
\alias{priorInfo<-}
\alias{priorInfo,DESeqResults-method}
\alias{priorInfo<-,DESeqResults,list-method}
\title{Accessors for the 'priorInfo' slot of a DESeqResults object.}
\usage{
priorInfo(object, ...)

priorInfo(object, ...) <- value

\S4method{priorInfo}{DESeqResults}(object)

\S4method{priorInfo}{DESeqResults,list}(object) <- value
}
\arguments{
\item{object}{a \code{DESeqResults} object}

\item{...}{additional arguments}

\item{value}{a \code{list}}
}
\description{
The priorInfo slot contains details about the prior on log fold changes
}
