\docType{methods}
\name{show,DESeqResults-method}
\alias{show,DESeqResults-method}
\title{Show method for DESeqResults objects}
\usage{
\S4method{show}{DESeqResults}(object)
}
\arguments{
  \item{object}{a DESeqResults object}
}
\description{
Prints out the information from the metadata columns of the
results object regarding the log2 fold changes and
p-values, then shows the DataFrame using the standard
method.
}

