% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,DESeqResults-method}
\title{Show method for DESeqResults objects}
\usage{
\S4method{show}{DESeqResults}(object)
}
\arguments{
\item{object}{a DESeqResults object}
}
\description{
Prints out the information from the metadata columns
of the results object regarding the log2 fold changes
and p-values, then shows the DataFrame using the
standard method.
}
\author{
Michael Love
}

