% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{sparsityPlot}
\alias{sparsityPlot}
\title{Sparsity plot}
\usage{
sparsityPlot(x, normalized = TRUE, ...)
}
\arguments{
\item{x}{a matrix or DESeqDataSet}

\item{normalized}{whether to normalize the counts from a DESeqDataSEt}

\item{...}{passed to \code{plot}}
}
\description{
Simple plot of the concentration of counts in a single sample over the
sum of counts per gene.
}
\examples{
dds <- makeExampleDESeqDataSet(n=1000,m=4,dispMeanRel=function(x) .5)
dds <- estimateSizeFactors(dds)
sparsityPlot(dds)
}

