% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summarizeResults}
\alias{summarizeResults}
\title{Summarize DESeq results}
\usage{
summarizeResults(res, alpha = 0.1)
}
\arguments{
\item{res}{a \code{\link{DESeqResults}} object}

\item{alpha}{the adjusted p-value cutoff}
}
\description{
Print a summary of the results from a DESeq analysis.
}
\examples{
example("DESeq")
summarizeResults(res)
}

