% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{unmix}
\alias{unmix}
\title{Unmix samples using loss in a variance stabilized space}
\usage{
unmix(x, pure, alpha, shift, loss = 1)
}
\arguments{
\item{x}{normalized counts or TPMs of the samples to be unmixed}

\item{pure}{normalized counts or TPMs of the "pure" samples}

\item{alpha}{for normalized counts, the dispersion of the data
when a negative binomial model is fit. this can be found by examining
the asymptotic value of \code{dispersionFunction(dds)}, when using
\code{fitType="parametric"} or the mean value when using
\code{fitType="mean"}.}

\item{shift}{for TPMs, the shift which approximately stabilizes the variance
of log shifted TPMs. Can be assessed with \code{vsn::meanSdPlot}.}

\item{loss}{either 1 (for L1) or 2 (for squared) loss function.
Default is 1.}
}
\value{
mixture components for each sample (rows), which sum to 1.
}
\description{
Unmixes samples in \code{x} according to \code{pure} components,
using numerical optimization. The components in \code{pure}
are added on the scale of gene expression (either normalized counts, or TPMs).
The loss function when comparing fitted expression to the
samples in \code{x} occurs in a variance stabilized space.
This task is sometimes referred to as "deconvolution",
and can be used, for example, to identify contributions from
various tissues.
Note: if the \code{pbapply} package is installed a progress bar
will be displayed while mixing components are fit.
}
