% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DESeqResults-class}
\alias{DESeqResults}
\alias{DESeqResults-class}
\title{DESeqResults object and constructor}
\usage{
DESeqResults(DataFrame)
}
\arguments{
\item{DataFrame}{a DataFrame of results, standard column names are:
baseMean, log2FoldChange, lfcSE, stat, pvalue, padj.}
}
\value{
a DESeqResults object
}
\description{
This constructor function would not typically be used by "end users".
This simple class extends the DataFrame class of the IRanges package
to allow other packages to write methods for results
objects from the DESeq2 package. It is used by \code{\link{results}}
to wrap up the results table.
}

