% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dispersionFunction}
\alias{dispersionFunction}
\alias{dispersionFunction,DESeqDataSet-method}
\alias{dispersionFunction.DESeqDataSet}
\alias{dispersionFunction<-}
\alias{dispersionFunction<-,DESeqDataSet,function-method}
\title{Accessors for the 'dispersionFunction' slot of a DESeqDataSet object.}
\usage{
dispersionFunction(object, ...)

dispersionFunction(object, ...) <- value

\S4method{dispersionFunction}{DESeqDataSet}(object)

\S4method{dispersionFunction}{DESeqDataSet,function}(object,estimateVar)<-value

\S4method{dispersionFunction}{DESeqDataSet}(object)
}
\arguments{
\item{object}{a \code{DESeqDataSet} object.}

\item{value}{a \code{function}}

\item{estimateVar}{whether to estimate the variance of dispersion residuals.
setting to FALSE is needed, e.g. within \code{estimateDispersionsMAP} when
called on a subset of the full dataset in parallel execution.}

\item{...}{additional arguments}
}
\description{
The dispersion function is calculated by \code{\link{estimateDispersions}} and
used by \code{\link{varianceStabilizingTransformation}}.  Parametric dispersion
fits store the coefficients of the fit as attributes in this slot.
}
\examples{
example("estimateDispersions")
dispersionFunction(dds)
}

