% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{dispersions}
\alias{dispersions}
\alias{dispersions,DESeqDataSet-method}
\alias{dispersions<-}
\alias{dispersions<-,DESeqDataSet,numeric-method}
\title{Accessor functions for the dispersion estimates in a DESeqDataSet
object.}
\usage{
dispersions(object, ...)

dispersions(object, ...) <- value

\S4method{dispersions}{DESeqDataSet}(object)

\S4method{dispersions}{DESeqDataSet,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{DESeqDataSet} object.}

\item{...}{additional arguments}

\item{value}{the dispersions to use for the Negative Binomial modeling}
}
\description{
The dispersions for each row of the DESeqDataSet.  Generally,
these are set by \code{\link{estimateDispersions}}.
}
\author{
Simon Anders
}
\seealso{
\code{\link{estimateDispersions}}
}

