% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{estimateSizeFactors}
\alias{estimateSizeFactors}
\alias{estimateSizeFactors,DESeqDataSet-method}
\alias{estimateSizeFactors.DESeqDataSet}
\title{Estimate the size factors for a DESeqDataSet}
\usage{
\S4method{estimateSizeFactors}{DESeqDataSet}(object,locfunc=median,geoMeans)

\S4method{estimateSizeFactors}{DESeqDataSet}(object, locfunc = median,
  geoMeans)
}
\arguments{
\item{object}{a DESeqDataSet}

\item{locfunc}{a function to compute a location for a sample. By default, the
median is used. However, especially for low counts, the
\code{\link[genefilter]{shorth}} function from the genefilter package may give better results.}

\item{geoMeans}{by default this is not provided and the
geometric means of the counts are calculated within the function.
A vector of geometric means from another count matrix can be provided
for a "frozen" size factor calculation}
}
\value{
The DESeqDataSet passed as parameters, with the size factors filled
in.
}
\description{
Estimate the size factors for a DESeqDataSet
}
\details{
This function estimates the size factors using the
"median ratio method" described by Equation 5 in Ander and Huber (2010).
The estimated size factors can be accessed using \code{\link{sizeFactors}}.
Alternative library size estimators can also be supplied
using \code{\link{sizeFactors}}.

Typically, the function is called with the idiom:

\code{dds <- estimateSizeFactors(dds)}

See \code{\link{DESeq}} for a description of the use of size factors in the GLM.
One should call this function after \code{\link{DESeqDataSet}}
unless size factors are manually specified with \code{\link{sizeFactors}}.
Alternatively, gene-specific normalization factors for each sample can be provided using
\code{\link{normalizationFactors}} which will always preempt \code{\link{sizeFactors}}
in calculations.

Internally, the function calls \code{\link{estimateSizeFactorsForMatrix}},
which provides more details on the calculation.
}
\examples{
dds <- makeExampleDESeqDataSet()
dds <- estimateSizeFactors(dds)
sizeFactors(dds)
geoMeans <- exp(rowMeans(log(counts(dds))))
dds <- estimateSizeFactors(dds,geoMeans=geoMeans)
sizeFactors(dds)
}
\author{
Simon Anders
}
\references{
\itemize{
  \item Simon Anders, Wolfgang Huber: Differential expression analysis for sequence count data. Genome Biology 11 (2010) R106, \url{http://dx.doi.org/10.1186/gb-2010-11-10-r106}
}
}
\seealso{
\code{\link{estimateSizeFactorsForMatrix}}
}

