% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{lfcShrink}
\alias{lfcShrink}
\title{Shrink log2 fold changes}
\usage{
lfcShrink(dds, coef, coefVar, res)
}
\arguments{
\item{dds}{a DESeqDataSet object, which has been run through
\code{\link{DESeq}}, or at the least, \code{\link{estimateDispersions}}}

\item{coef}{the number of the coefficient (LFC) to shrink,
consult \code{resultsNames(dds)} after running \code{DESeq(dds, betaPrior=FALSE)}}

\item{coefVar}{the variance (width) of the prior to use in shrinking the LFC}

\item{res}{a DESeqResults object (can be missing)}
}
\value{
if \code{res} is not missing, a DESeqResults object with
the \code{log2FoldChange} column replaced with a shrunken LFC.
If \code{res} is missing, just the shrunken LFC vector.
}
\description{
This function adds shrunken log2 fold changes (LFC) to a
results table which was run without LFC moderation.
Note: this function is still being prototyped.
}
\examples{

 dds <- makeExampleDESeqDataSet(betaSD=1)
 dds <- DESeq(dds, betaPrior=FALSE)
 res <- results(dds)
 res.shr <- lfcShrink(dds, coef=2, coefVar=1, res)

}

