% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/core.R
\name{nbinomLRT}
\alias{nbinomLRT}
\title{Likelihood ratio test (chi-squared test) for GLMs}
\usage{
nbinomLRT(object, full = design(object), reduced, betaPrior = FALSE,
  betaPriorVar, modelMatrixType = "standard", maxit = 100,
  useOptim = TRUE, quiet = FALSE, useQR = TRUE)
}
\arguments{
\item{object}{a DESeqDataSet}

\item{full}{the full model formula, this should be the formula in
\code{design(object)}.
alternatively, can be a matrix}

\item{reduced}{a reduced formula to compare against, e.g.
the full model with a term or terms of interest removed.
alternatively, can be a matrix}

\item{betaPrior}{whether or not to put a zero-mean normal prior on
the non-intercept coefficients
While the beta prior is used typically, for the Wald test, it can
also be specified for the likelihood ratio test. For more details
on the calculation, see \code{\link{nbinomWaldTest}}.}

\item{betaPriorVar}{a vector with length equal to the number of
model terms including the intercept.
which if missing is estimated from the rows which do not have any
zeros}

\item{modelMatrixType}{either "standard" or "expanded", which describe
how the model matrix, X of the formula in \code{\link{DESeq}}, is
formed. "standard" is as created by \code{model.matrix} using the
design formula. "expanded" includes an indicator variable for each
level of factors in addition to an intercept,
in order to ensure that the log2 fold changes are independent
of the choice of reference level.
betaPrior must be set to TRUE in order for expanded model matrices
to be fit.}

\item{maxit}{the maximum number of iterations to allow for convergence of the
coefficient vector}

\item{useOptim}{whether to use the native optim function on rows which do not
converge within maxit}

\item{quiet}{whether to print messages at each step}

\item{useQR}{whether to use the QR decomposition on the design
matrix X while fitting the GLM}
}
\value{
a DESeqDataSet with new results columns accessible
with the \code{\link{results}} function.  The coefficients and standard errors are
reported on a log2 scale.
}
\description{
This function tests for significance of change in deviance between a
full and reduced model which are provided as \code{formula}.
Fitting uses previously calculated \code{\link{sizeFactors}} (or \code{\link{normalizationFactors}})
and dispersion estimates.
}
\details{
The difference in deviance is compared to a chi-squared distribution
with df = (reduced residual degrees of freedom - full residual degrees of freedom).
This function is comparable to the \code{nbinomGLMTest} of the previous version of DESeq
and an alternative to the default \code{\link{nbinomWaldTest}}.
}
\examples{
dds <- makeExampleDESeqDataSet()
dds <- estimateSizeFactors(dds)
dds <- estimateDispersions(dds)
dds <- nbinomLRT(dds, reduced = ~ 1)
res <- results(dds)
}
\seealso{
\code{\link{DESeq}}, \code{\link{nbinomWaldTest}}
}

