% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotCounts}
\alias{plotCounts}
\title{Plot of normalized counts for a single gene on log scale}
\usage{
plotCounts(dds, gene, intgroup = "condition", normalized = TRUE,
  transform = FALSE, ...)
}
\arguments{
\item{dds}{a \code{DESeqDataSet}}

\item{gene}{a character, specifying the name of the gene to plot}

\item{intgroup}{a character vector of names in \code{colData(x)} to use for grouping}

\item{normalized}{whether the counts should be normalized by size factor
(default is TRUE)}

\item{transform}{whether to present log2 counts (TRUE) or
to present the counts on the log scale (FALSE, default)}

\item{...}{arguments passed to plot}
}
\description{
Plot of normalized counts for a single gene on log scale
}
\examples{
dds <- makeExampleDESeqDataSet()
plotCounts(dds, "gene1")
}

