% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotPCA}
\alias{plotPCA}
\title{Sample PCA plot from variance-stabilized data}
\usage{
plotPCA(x, intgroup = "condition", ntop = 500, col)
}
\arguments{
\item{x}{a SummarizedExperiment, with data in \code{assay(x)},
produced for example by either \code{\link{varianceStabilizingTransformation}}
or \code{\link{rlogTransformation}}}

\item{intgroup}{interesting groups: a character vector of names in \code{colData(x)} to use for grouping}

\item{ntop}{number of top genes to use for principal components, selected by highest
row variance}

\item{col}{a vector of colors for each level of intgroup}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
This plot helps to check for batch effects and the like.
}
\note{
See the vignette for an example of variance stabilization and PCA plots.
Note that the source code of plotPCA is very simple and commented.
Users should find it easy to customize this function.
}
\examples{
dds <- makeExampleDESeqDataSet(betaSD=1)
rld <- rlog(dds)
plotPCA(rld)
}
\author{
Wolfgang Huber
}

