% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary.DESeqResults}
\title{Summarize DESeq results}
\usage{
\method{summary}{DESeqResults}(object, alpha=.1, \dots)
}
\arguments{
\item{object}{a \code{\link{DESeqResults}} object}

\item{alpha}{the adjusted p-value cutoff}

\item{...}{additional arguments}
}
\description{
Print a summary of the results from a DESeq analysis.
}
\examples{
example("DESeq")
summary(res)
}
\author{
Michael Love
}

